api_yamdb/requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
asgiref==3.2.10
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
install==1.3.4
isort==5.9.3
mccabe==0.6.1
packaging==21.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
style==1.1.0
toml==0.10.2
update==0.0.1
urllib3==1.26.7
python-dotenv==0.13.0
